style: moved get_random_generator to utils
Also refactored utils to be a folder to avoid circular import: utils.py
imports space.py, and I need get_random_generator in space.py, so I
can't define it in utils.py. By creating the utils folder, moving
utils.py into it, and importing every function in __init__.py, I keep
the functionality. I can then define get_random_generartor in
get_rng.py in the utils folder, and import that into __init__.py as
well.
